화면 렌더링과 로직 분리
