+++ src/test/java/game/CarTest.java
@@ -7,14 +7,6 @@ import org.junit.jupiter.api.Test;
 import static org.assertj.core.api.Assertions.*;
 
 public class CarTest {
-    @Test
-    @DisplayName("자동차의 현재 위치를 나타내는 문자열 생성")
-    public void currentPosition() {
-        Car mySummerCar = new Car(5);
-
-        assertThat(mySummerCar.getMoveTrajectory())
-                .isEqualTo("-----");
-    }
 
     @Test
     @DisplayName("자동차 움직이기 테스트")
